changed type from double to int

+++ tasks/levonychev_i_min_val_rows_matrix/common/include/common.hpp
@@ -8,8 +8,8 @@
 
 namespace levonychev_i_min_val_rows_matrix {
 
-using InType = std::tuple<std::vector<double>, size_t, size_t>;
-using OutType = std::vector<double>;
+using InType = std::tuple<std::vector<int>, size_t, size_t>;
+using OutType = std::vector<int>;
 using TestType = std::tuple<size_t, size_t>;
 using BaseTask = ppc::task::Task<InType, OutType>;
